+++ .browserslistrc
@@ -1,0 +1,3 @@
+last 2 versions
+not dead
+>0.5%
