--- index.html
@@ -1,25 +1,0 @@
-<!doctype html>
-		<html lang="ru">
-			<head>
-				<meta charset="UTF-8">
-				<meta name="viewport" content="width=device-width, initial-scale=1.0">
-				<title>Карусель</title>
-				<link rel="stylesheet" href="main.css">
-				
-			</head>
-			<body>
-				<div id="container" class="galleryContainer">
-					<div class="volumeElem"></div>
-					<div class="element" style="background-color: red;"></div>
-					<div class="element" style="background-color: blue;"></div>
-					<div class="element" style="background-color: red;"></div>
-					<div class="element" style="background-color: blue;"></div>
-					<div class="element" style="background-color: red;"></div>
-					<div class="element" style="background-color: blue;"></div>
-					<div class="element" style="background-color: red;"></div>
-					
-				</div>
-				<script src="dist/main.js"></script>
-				
-			</body>
-		</html>
