+++ package.json
@@ -1,24 +1,30 @@
 {
-  "name": "webpack-5-lesson",
+  "name": "carousel-gallery",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "webpack serve --mode development",
-    "build": "webpack --mode production"
+    "start": "webpack serve",
+    "build-dev": "webpack",
+    "build-prod": "webpack --node-env=production",
+    "clear": "rd /s /q dist"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Kirillka922/carousel-gallery.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.17.3",
-    "babel-loader": "^8.2.3",
-    "mini-css-extract-plugin": "^2.5.3",
-    "webpack": "^5.69.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
+  "bugs": {
+    "url": "https://github.com/Kirillka922/carousel-gallery/issues"
   },
-  "dependencies": {
-    "@babel/preset-env": "^7.21.5"
+  "homepage": "https://github.com/Kirillka922/carousel-gallery#readme",
+  "devDependencies": {
+    "html-loader": "^4.2.0",
+    "html-webpack-plugin": "^5.5.3",
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
