+++ src/index.html
@@ -1,0 +1,24 @@
+
+		<!DOCTYPE html>
+<html>
+	
+  <head>
+    <meta charset="UTF-8" />
+	<title>hello world</title>
+	<link rel="stylesheet" href="main.css">
+  </head>
+  <body>
+
+    <div id="container" class="galleryContainer">
+      <div class="volumeElem"></div>
+      <div class="element" style="background-color: red;"></div>
+      <div class="element" style="background-color: blue;"></div>
+      <div class="element" style="background-color: red;"></div>
+      <div class="element" style="background-color: blue;"></div>
+      <div class="element" style="background-color: red;"></div>
+      <div class="element" style="background-color: blue;"></div>
+      <div class="element" style="background-color: red;"></div>
+      
+    </div>
+  </body>
+</html>
