+++ webpack.config.js
@@ -1,14 +1,35 @@
-let path = require("path");
+const path = require("path");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const mode = process.env.NODE_ENV || "development";
+const devMode = mode === "development";
+const target = devMode ? "web" : "browserslist";
+const devtool = devMode ? "source-map" : undefined;
 
-let conf = {
-  entry: "./src/main.js",
+module.exports = {
+  mode,
+  target,
+  devtool,
+  entry: path.resolve(__dirname, "src", "index.js"),
   output: {
-    path: path.resolve(__dirname, "./dist"),
-    filename: "main.js",
-    //publicPath: "/dist/",
+    path: path.resolve(__dirname, "dist"),
+    clean: true,
+    filename: "index.[contenthash].js",
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: path.resolve(__dirname, "src", "index.html"),
+    }),
+  ],
+  module: {
+    rules: [
+      {
+        test: /\.html$/i,
+        loader: "html-loader",
+      },
+    ],
   },
   devServer: {
-    //port: 3031,
+    port: 3031,
     proxy: {
       "/": {
         target: "https://966e3a17bd14-817885256656399495.ngrok-free.app",
@@ -17,9 +38,4 @@ let conf = {
       },
     },
   },
-};
-module.exports = (env, options) => {
-  let isProd = options.mode === "production";
-  conf.devtool = isProd ? false : "eval-cheap-module-source-map";
-  return conf;
 };
